% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_functions.R
\name{calculate_aggregated_value}
\alias{calculate_aggregated_value}
\title{Calculate one of the aggregating statistic.}
\usage{
calculate_aggregated_value(dataset, column, agg_fun, as_num = FALSE)
}
\arguments{
\item{dataset}{survey dataset}

\item{column}{column name used for calculating statistic}

\item{agg_fun}{function used to calculate statistic - one of \code{mean}, \code{min}, \code{max} or \code{median}}

\item{as_num}{if \code{TRUE} returns value as numeric otherwise returns data.frame object (default \code{FALSE})}
}
\value{
value of calculated statistic
}
\description{
For selected \code{column} calculate \code{agg_fun}.
}
\examples{
calculate_aggregated_value(survey, "NoChild", "mean")
calculate_aggregated_value(survey, "NoTeen", "min", TRUE)
calculate_aggregated_value(survey, "BMI", "max", TRUE)
}
