% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_functions.R
\name{count_observations}
\alias{count_observations}
\title{Count number of observations in selected groups.}
\usage{
count_observations(dataset, grouping_cols)
}
\arguments{
\item{dataset}{survey dataset}

\item{grouping_cols}{column name or vector of column names used for creating groups}
}
\value{
data frame with counts
}
\description{
Dataset is grouped by \code{grouping_cols} and for each combination of values
number of observations in each group is calculated.
}
\examples{
count_observations(survey, "Gender")
count_observations(survey, c("Gender", "Age"))

}
