% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_functions.R
\name{draw_activity_effect}
\alias{draw_activity_effect}
\title{Draw activity effect}
\usage{
draw_activity_effect(dataset, exercise_type, effect_on_col, corr = TRUE)
}
\arguments{
\item{dataset}{survey dataset}

\item{exercise_type}{a}

\item{effect_on_col}{a}

\item{corr}{a}
}
\value{
ggplot
}
\description{
Description
}
\examples{
draw_activity_effect(survey, "Vigorous Work", "Weight", FALSE)
draw_activity_effect(survey, "Moderate Work", "BMI")
}
