% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_functions.R
\name{draw_barplot}
\alias{draw_barplot}
\title{Draw barplot}
\usage{
draw_barplot(dataset, counted_col, grouping_col)
}
\arguments{
\item{dataset}{survey dataset}

\item{counted_col}{a}

\item{grouping_col}{a}
}
\value{
ggplot
}
\description{
Description
}
\examples{
draw_barplot(survey, "Age", "Ethnicity")
draw_barplot(survey, "Income", "Gender")
}
