% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_functions.R
\name{draw_exercise_time_density}
\alias{draw_exercise_time_density}
\title{Draw density}
\usage{
draw_exercise_time_density(dataset, exercise_type, grouping_col)
}
\arguments{
\item{dataset}{survey dataset}

\item{exercise_type}{a}

\item{grouping_col}{a}
}
\value{
ggplot
}
\description{
Description
}
\examples{
draw_exercise_time_density(survey, "Vigorous Work", "Gender")
draw_exercise_time_density(survey, "Walking/Biking", "Age")
}
