% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploreNHANES_package.R
\docType{package}
\name{exploreNHANES-package}
\alias{exploreNHANES-package}
\alias{_PACKAGE}
\alias{exploreNHANES}
\title{Explore NHANES data}
\description{
\pkg{exploreNHANES} is the \R package created to explore a subset of NHANES
publication, drawing attention to the prevalence of exercise and associated
health effects across different demographics. The functions provided with
a package allow to explore and visualize data.
}
\references{
\emph{\pkg{exploreNHANES} Package homepage}, \url{https://github.com/jjankowiak/exploreNHANES}
}
\author{
Justyna Jankowiak
}
